# openapi.yaml
openapi: 3.0.3
info:
  title: Matchmaking and Notification Service API
  version: 1.0.0
  description: |
    API-first definition for the Matchmaking and Notification microservice.
    Each resource (matches, offers, health) includes GET, POST, PUT, and DELETE methods.
    These endpoints are defined for design purposes only — no implementation is required.

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /matches:
    get:
      summary: Get all matches
      description: Retrieve a list of all matches.
      responses:
        '200':
          description: A list of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Match'

    post:
      summary: Create a new match
      description: Add a new donor-organ match.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Match'
      responses:
        '201':
          description: Match created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'

    put:
      summary: Update all matches (bulk operation placeholder)
      description: Update multiple matches (for demonstration purposes).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Match'
      responses:
        '200':
          description: Matches updated successfully

    delete:
      summary: Delete all matches
      description: Delete all matches (for demonstration purposes).
      responses:
        '204':
          description: All matches deleted

  /matches/{id}:
    get:
      summary: Get a match by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Match retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Match'
        '404':
          description: Match not found

    put:
      summary: Update a specific match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Match'
      responses:
        '200':
          description: Match updated successfully
        '404':
          description: Match not found

    post:
      summary: Duplicate or reprocess a specific match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Match reprocessed successfully

    delete:
      summary: Delete a specific match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Match deleted successfully
        '404':
          description: Match not found

  /offers:
    get:
      summary: Get all offers
      description: Retrieve a list of all offers.
      responses:
        '200':
          description: A list of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Offer'

    post:
      summary: Create a new offer
      description: Add a new allocation offer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Offer'
      responses:
        '201':
          description: Offer created successfully

    put:
      summary: Update multiple offers (bulk operation placeholder)
      description: Update several offers at once (for demonstration purposes).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/Offer'
      responses:
        '200':
          description: Offers updated successfully

    delete:
      summary: Delete all offers
      description: Delete all offers (for demonstration purposes).
      responses:
        '204':
          description: All offers deleted

  /offers/{id}:
    get:
      summary: Get an offer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Offer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
        '404':
          description: Offer not found

    put:
      summary: Update an offer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Offer'
      responses:
        '200':
          description: Offer updated successfully
        '404':
          description: Offer not found

    post:
      summary: Duplicate or resend a specific offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Offer duplicated successfully

    delete:
      summary: Delete an offer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Offer deleted successfully
        '404':
          description: Offer not found

  /health:
    get:
      summary: Health check
      description: Retrieve API health status.
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok

    post:
      summary: Report custom health status
      description: Send custom health status for testing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        '201':
          description: Health status recorded

    put:
      summary: Update health configuration
      description: Update simulated health configuration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mode:
                  type: string
      responses:
        '200':
          description: Health configuration updated

    delete:
      summary: Reset health data
      description: Delete or reset simulated health state.
      responses:
        '204':
          description: Health data reset successfully

components:
  schemas:
    Match:
      type: object
      properties:
        id:
          type: string
          example: "match123"
        donorId:
          type: string
          example: "donor789"
        organId:
          type: string
          example: "organ456"
        status:
          type: string
          enum: [pending, accepted, declined, expired]
          example: "pending"
      required:
        - donorId
        - organId

    Offer:
      type: object
      properties:
        id:
          type: string
          example: "offer001"
        matchId:
          type: string
          example: "match123"
        status:
          type: string
          enum: [pending, accepted, declined, expired]
          example: "pending"
      required:
        - matchId